Remove old drag-drop files
