+++ client/ListMapper/list-mapper/src/resources/elements/drag-drop/leveling/drag-drop-ca.html
@@ -1,9 +1,7 @@
 <template>
   <require from='../../simple-list'></require>
-  <require from='../../../attributes/enable-drag-drop'></require>
   
   <simple-list 
-    enable-drag-drop 
     view-model.ref="simpleListVM"
     list-data.bind="listData"
   ></simple-list>
