+++ client/ListMapper/list-mapper/src/resources/elements/drag-drop/leveling/drag-drop-v1-ce.html
@@ -1,10 +1,10 @@
 <template>
   <require from='../../simple-list'></require>
-  <require from='../../../attributes/drag-drop-v1'></require>
+  <require from='../../../attributes/enable-drag-drop'></require>
 
   <div class='drag-drop-v1-ce'>
-    <simple-list drag-drop-v1  list-data.bind='listData'></simple-list>
-    <simple-list drag-drop-v1  list-data.bind='listDataCommands'></simple-list>
+    <simple-list enable-drag-drop  list-data.bind='listData'></simple-list>
+    <simple-list enable-drag-drop  list-data.bind='listDataCommands'></simple-list>
   </div>
 
 </template>
